requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# API & Web
requests>=2.31.0
websockets>=11.0
streamlit>=1.28.0

# Statistics
scipy>=1.11.0
statsmodels>=0.14.0

# Config & Utils
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0.0

# Monitoring
python-telegram-bot>=20.0
plotly>=5.15.0

# Testing
pytest>=7.4.0